Fixed cosine similarity
Previously, the model was looping comparisons between feature_matrix.iloc[start:end,:].

This meant that it begins with rows 0 to 999, which included the input song at 0, but then goes to 1000 to 1999, which means that the input song is not being compared against anymore.

Now, the input song has been isolated prior to the loop, and will be added into the comparison 1000 songs df every loop to ensure that it is the subject of comparison.